requirements.txt
BiblioPixel
BiblioPixelAnimations
pynetworktables
web.py

